cleaner: Stage B — extract ocr_render.py + text_surface_metrics.py from phase_clean.py
Per the cleaner-integration plan's Stage B + the cleaner correctness axes
memory rule "latest changes are canonical".

## What's extracted

11 functions moved from `src/glossapi/corpus/phase_clean.py` into the new
`src/glossapi/corpus/ocr_render.py`:

- _gap_has_at_most_n_nonwhitespace_chars
- _clean_fill_for_removed_span
- _merge_labeled_raw_spans
- _summarize_merged_labeled_spans
- _render_page_from_merged_labeled_spans
- _render_page_with_labeled_spans_result
- _render_page_with_labeled_spans
- _annotate_page_with_labeled_spans
- _utf8_prefix_byte_offsets
- _span_repeat_count
- _build_match_index_rows

These collectively own the analyzer/renderer separation: phase_clean.py
decides WHAT spans exist; ocr_render.py renders HOW those spans become
page text and debug sidecars.

## Body resolution

Per the function-by-function comparison done in the cleaner-integration
audit, 5 functions were EXACT match vs faa1362 and 6 were DIFF. Per the
"latest changes canonical" rule, this PR uses **dev's bodies for all 11**
(dev's Apr 14 OCR speedup wave is later than faa1362's Apr 12 work).

The faa1362-only helper `_build_debug_match_open_tag` is NOT brought
over because dev's `_render_page_from_merged_labeled_spans` (73 lines vs
faa1362's 44) inlines the equivalent logic and doesn't need the helper.

## Also added

`src/glossapi/corpus/text_surface_metrics.py` (48 lines, from faa1362
verbatim — new module): `sanitized_char_count` + `_strip_latex_envs_for_char_count`.
Shared "published-surface" metric helpers used by export-facing
metadata refresh.

## phase_clean.py changes

- 11 function definitions removed (~330 lines).
- New imports: `from .ocr_render import (...)` (11 names) and
  `from .text_surface_metrics import sanitized_char_count`.
- Net: phase_clean.py 4929 → 4597 lines.

## docs/architecture/ocr_cleaning_runtime.md

Padded from 118 → 186 lines with the previously-missing sections:
- Code Layout (now accurate — describes the modules that exist)
- Stage Boundary: clean_ocr() vs clean()
- Field Ownership (OCR-owned vs clean/export-owned parquet fields)

## Verification

- python3 ast.parse: ocr_render.py OK, text_surface_metrics.py OK,
  phase_clean.py OK.
- Direct module import + sanitized_char_count smoke: works correctly.
- pytest tests/test_corpus_clean_enhancements.py: 67 passed, 2 failed.
  Both failures (test_clean_flags_uppercase_glyph_noise and
  test_clean_token_category_debug_exports_synthetic_pages) are
  PRE-EXISTING dev failures — verified by running the same tests against
  unmodified origin/development. NOT introduced by this extraction.
